# backend/src/plants/routers/tracking.py
"""Router for plant tracking and photo endpoints."""

from datetime import datetime
from typing import List, Optional

from fastapi import APIRouter, Depends, File, Form, UploadFile, status
from sqlalchemy.orm import Session

from src.auth.models import User
from src.database.session import get_db
from src.plants.dependencies import get_current_user_from_request
from src.plants.schemas import PlantPhotoResponse

router = APIRouter(prefix="/plants/track", tags=["plant-tracking"])


@router.post(
    "/photos", response_model=PlantPhotoResponse, status_code=status.HTTP_201_CREATED
)
async def upload_plant_photo(
    plant_id: int = Form(...),
    description: Optional[str] = Form(None),
    file: UploadFile = File(...),
    db: Session = Depends(get_db),
    current_user: User = Depends(get_current_user_from_request),
):
    """Upload a new photo for plant tracking."""
    # TODO: Implement photo upload with:
    # - File validation and processing
    # - AI analysis for growth insights
    # - Storage in S3 with metadata
    # - Database record creation

    return PlantPhotoResponse(
        id=1,
        plant_id=plant_id,
        user_id=current_user.id,
        filename=file.filename or "unknown.jpg",
        original_filename=file.filename,
        file_size=0,  # Would be calculated
        content_type=file.content_type or "image/jpeg",
        description=description,
        ai_analysis=None,
        created_at=datetime.now(),
        url="https://placeholder.com/photo.jpg",
        thumbnail_url="https://placeholder.com/thumb.jpg",
    )


@router.get("/photos/{plant_id}", response_model=List[PlantPhotoResponse])
async def get_plant_photos(
    plant_id: int,
    db: Session = Depends(get_db),
    current_user: User = Depends(get_current_user_from_request),
):
    """Get all photos for a specific plant."""
    # TODO: Implement photo retrieval
    return []


@router.get("/progress/{plant_id}")
async def get_plant_progress(
    plant_id: int,
    db: Session = Depends(get_db),
    current_user: User = Depends(get_current_user_from_request),
):
    """Get AI-generated progress insights from photo history."""
    # TODO: Implement progress analysis
    return {
        "insights": [
            "Growth +15% in height from last month",
            "Leaves appear greener with RGB delta +20",
            "Potential new budding detected in latest photo",
        ],
        "metrics": {
            "height_change": "15%",
            "leaf_color_improvement": "20%",
            "new_growth_detected": True,
        },
        "recommendations": [
            "Continue current care routine",
            "Consider providing support for new growth",
            "Monitor for flowering in coming weeks",
        ],
    }