# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "plant-assistant",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/plants/reminders/": {
      "post": {
        "tags": [
          "plant-reminders"
        ],
        "summary": "Create Reminder",
        "description": "Create a new reminder for a plant.",
        "operationId": "create_reminder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "plant-reminders"
        ],
        "summary": "Get Reminders",
        "description": "Get paginated list of user's reminders.",
        "operationId": "get_reminders",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          },
          {
            "name": "plant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          },
          {
            "name": "reminder_type",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReminderType"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReminderPriority"
            }
          },
          {
            "name": "is_completed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Is Completed"
            }
          },
          {
            "name": "overdue_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Overdue Only"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "next_due_date",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc",
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plants/reminders/upcoming": {
      "get": {
        "tags": [
          "plant-reminders"
        ],
        "summary": "Get Upcoming Reminders",
        "description": "Get upcoming reminders for the next N days.",
        "operationId": "get_upcoming_reminders",
        "parameters": [
          {
            "name": "days_ahead",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 7,
              "title": "Days Ahead"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReminderResponse"
                  },
                  "title": "Response Plant-Reminders-Get Upcoming Reminders"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plants/reminders/{reminder_id}": {
      "get": {
        "tags": [
          "plant-reminders"
        ],
        "summary": "Get Reminder",
        "description": "Get a specific reminder by ID.",
        "operationId": "get_reminder",
        "parameters": [
          {
            "name": "reminder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reminder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "plant-reminders"
        ],
        "summary": "Update Reminder",
        "description": "Update a reminder.",
        "operationId": "update_reminder",
        "parameters": [
          {
            "name": "reminder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reminder Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "plant-reminders"
        ],
        "summary": "Delete Reminder",
        "description": "Delete a reminder.",
        "operationId": "delete_reminder",
        "parameters": [
          {
            "name": "reminder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reminder Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plants/track/photos": {
      "post": {
        "tags": [
          "plant-tracking"
        ],
        "summary": "Upload Plant Photo",
        "description": "Upload a new photo for plant tracking.",
        "operationId": "upload_plant_photo",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_plant-tracking-upload_plant_photo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantPhotoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plants/track/photos/{plant_id}": {
      "get": {
        "tags": [
          "plant-tracking"
        ],
        "summary": "Get Plant Photos",
        "description": "Get all photos for a specific plant.",
        "operationId": "get_plant_photos",
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlantPhotoResponse"
                  },
                  "title": "Response Plant-Tracking-Get Plant Photos"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plants/track/progress/{plant_id}": {
      "get": {
        "tags": [
          "plant-tracking"
        ],
        "summary": "Get Plant Progress",
        "description": "Get AI-generated progress insights from photo history.",
        "operationId": "get_plant_progress",
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_plant-tracking-upload_plant_photo": {
        "properties": {
          "plant_id": {
            "type": "integer",
            "title": "Plant Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "plant_id",
          "file"
        ],
        "title": "Body_plant-tracking-upload_plant_photo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlantPhotoResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 255,
            "title": "Filename"
          },
          "original_filename": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Filename"
          },
          "file_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "File Size"
          },
          "content_type": {
            "type": "string",
            "maxLength": 50,
            "title": "Content Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "plant_id": {
            "type": "integer",
            "title": "Plant Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "ai_analysis": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Analysis"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "file_size",
          "content_type",
          "id",
          "plant_id",
          "user_id",
          "ai_analysis",
          "created_at",
          "url"
        ],
        "title": "PlantPhotoResponse",
        "description": "Schema for plant photo response."
      },
...
      "ReminderCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "reminder_type": {
            "$ref": "#/components/schemas/ReminderType"
          },
          "priority": {
            "$ref": "#/components/schemas/ReminderPriority",
            "default": "medium"
          },
          "cron_expression": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cron Expression",
            "description": "Cron expression for recurring reminders"
          },
          "next_due_date": {
            "type": "string",
            "format": "date-time",
            "title": "Next Due Date"
          },
          "is_recurring": {
            "type": "boolean",
            "title": "Is Recurring",
            "default": false
          },
          "plant_id": {
            "type": "integer",
            "title": "Plant Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "reminder_type",
          "next_due_date",
          "plant_id"
        ],
        "title": "ReminderCreate",
        "description": "Schema for creating a new reminder."
      },
      "ReminderListResponse": {
        "properties": {
          "reminders": {
            "items": {
              "$ref": "#/components/schemas/ReminderResponse"
            },
            "type": "array",
            "title": "Reminders"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "reminders",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "ReminderListResponse",
        "description": "Schema for paginated reminder list response."
      },
      "ReminderPriority": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "ReminderPriority",
        "description": "Priority levels for reminders."
      },
      "ReminderResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "reminder_type": {
            "$ref": "#/components/schemas/ReminderType"
          },
          "priority": {
            "$ref": "#/components/schemas/ReminderPriority",
            "default": "medium"
          },
          "cron_expression": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cron Expression",
            "description": "Cron expression for recurring reminders"
          },
          "next_due_date": {
            "type": "string",
            "format": "date-time",
            "title": "Next Due Date"
          },
          "is_recurring": {
            "type": "boolean",
            "title": "Is Recurring",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "plant_id": {
            "type": "integer",
            "title": "Plant Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "is_completed": {
            "type": "boolean",
            "title": "Is Completed"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "plant_nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Nickname"
          },
          "is_overdue": {
            "type": "boolean",
            "title": "Is Overdue",
            "default": false
          },
          "days_until_due": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Until Due"
          }
        },
        "type": "object",
        "required": [
          "title",
          "reminder_type",
          "next_due_date",
          "id",
          "plant_id",
          "user_id",
          "is_completed",
          "completed_at",
          "created_at",
          "updated_at"
        ],
        "title": "ReminderResponse",
        "description": "Schema for reminder response."
      },
      "ReminderType": {
        "type": "string",
        "enum": [
          "watering",
          "fertilizing",
          "pruning",
          "repotting",
          "pest_check",
          "custom"
        ],
        "title": "ReminderType",
        "description": "Types of plant care reminders."
      },
      "ReminderUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "reminder_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReminderType"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReminderPriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "cron_expression": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cron Expression"
          },
          "next_due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Due Date"
          },
          "is_recurring": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Recurring"
          },
          "is_completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Completed"
          }
        },
        "type": "object",
        "title": "ReminderUpdate",
        "description": "Schema for updating a reminder."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}